# .futurxlab/agent-platform-api-spec.yaml
openapi: 3.0.0
info:
  title: Agent Content Generation Platform API
  version: 1.0.0
  description: Internal API for AI-powered content generation using Dify agents
  contact:
    name: FuturX Development Team
    email: dev@futurx.com
servers:
  - url: http://localhost:7860/api
    description: Development server
  - url: https://agent-platform.internal/api
    description: Production server
...
paths:
...
  /tasks:
    get:
      tags:
        - Tasks
      summary: List all task templates
      operationId: listTasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'

    post:
      tags:
        - Tasks
      summary: Create new task template
      operationId: createTask
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequestError'
  /tasks/{taskId}:
    get:
      tags:
        - Tasks
      summary: Get task details
      operationId: getTask
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/taskId'
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags:
        - Tasks
      summary: Update task template
      operationId: updateTask
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/taskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'

    delete:
      tags:
        - Tasks
      summary: Delete task template
      operationId: deleteTask
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/taskId'
      responses:
        '204':
          description: Task deleted successfully

  # Content Generation Endpoints (#REF-GEN-01)
...
  /history:
    get:
      tags:
        - History
      summary: List generation history
      operationId: listHistory
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: History list
          content:
            application/json:
              schema:
                type: object
                properties:
                  history:
                    type: array
                    items:
                      $ref: '#/components/schemas/HistoryItem'
                  total:
                    type: integer
                  hasMore:
                    type: boolean
  /history/{historyId}:
    get:
      tags:
        - History
      summary: Get history item details
      operationId: getHistoryItem
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/historyId'
      responses:
        '200':
          description: History item details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryDetail'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /history/{historyId}/export:
    get:
      tags:
        - History
      summary: Export history item as markdown
      operationId: exportHistory
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/historyId'
      responses:
        '200':
          description: Markdown export
          content:
            text/markdown:
              schema:
                type: string
        '404':
          $ref: '#/components/responses/NotFoundError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    taskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
      description: Task identifier
    historyId:
      name: historyId
      in: path
      required: true
      schema:
        type: string
      description: History item identifier
  schemas:
...
    Task:
      type: object
      required:
        - id
        - name
        - promptTemplate
        - createdAt
      properties:
        id:
          type: string
          example: "task_456"
        name:
          type: string
          example: "Daily Market Analysis"
        promptTemplate:
          type: string
          example: "Analyze the following market data: {input_text}"
        placeholders:
          type: array
          items:
            type: string
          example: ["input_text", "file_content"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TaskCreate:
      type: object
      required:
        - name
        - promptTemplate
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        promptTemplate:
          type: string
          minLength: 1
          maxLength: 10000
    TaskUpdate:
      type: object
      properties:
        name:
          type: string
        promptTemplate:
          type: string
    HistoryItem:
      type: object
      required:
        - id
        - agentName
        - taskName
        - createdAt
        - summary
      properties:
        id:
          type: string
        agentName:
          type: string
        taskName:
          type: string
        createdAt:
          type: string
          format: date-time
        summary:
          type: string
          description: First 200 characters of generated content
    HistoryDetail:
      allOf:
        - $ref: '#/components/schemas/HistoryItem'
        - type: object
          required:
            - fullContent
            - input
          properties:
            fullContent:
              type: string
              description: Complete generated content
            input:
              type: object
              properties:
                text:
                  type: string
                fileName:
                  type: string
            duration:
              type: integer
              description: Generation time in seconds
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
...
    BadRequestError:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "bad_request"
            message: "Validation failed"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "not_found"
            message: "Resource does not exist"
...